ml/requirements.txt
scikit-learn>=1.3
pandas>=2.1
numpy>=1.24
psycopg2-binary>=2.9
python-dotenv>=1.0
